refactor: move models into their respective modules, adjust project structure
