+++ provider/vndb/stats.go
@@ -2,17 +2,15 @@ package vndb
 
 import (
 	"encoding/json"
-
-	vndbmodels "kurohelper/models/vndb"
 )
 
-func GetStats() (*vndbmodels.Stats, error) {
+func GetStats() (*Stats, error) {
 	r, err := sendGetRequest("/stats")
 	if err != nil {
 		return nil, err
 	}
 
-	var res vndbmodels.Stats
+	var res Stats
 	err = json.Unmarshal(r, &res)
 	if err != nil {
 		return nil, err
